refactor: controller, view 폴더 구성
