+++ src/main/java/racingcar/view/RaceProgressView.java
@@ -1,6 +1,7 @@
-package racingcar;
+package racingcar.view;
 
 import java.util.List;
+import racingcar.RacingCar;
 
 public class RaceProgressView {
