+++ src/test/java/racingcar/RandomRaceControllerTest.java
@@ -6,6 +6,7 @@ import java.util.ArrayList;
 import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import racingcar.controller.RandomRaceController;
 
 class RandomRaceControllerTest {
     private RandomRaceController randomRaceController;
